requirements.txt
flask
pymongo
python-dotenv
pandas
numpy
scikit-learn
tqdm